Updated version and added enumeration type handler
In the `IceCoffee.Db4Net.csproj` file, the version number is updated from `1.0.6` to `1.0.7`.

In the `EnumStringTypeHandler.cs` file, a new Dapper type handler `EnumStringTypeHandler<T>` is added to convert enumeration types to strings for storage in the TEXT column of the database. This handler implements the `Parse` and `SetValue` methods, supports use in .NET Core and other frameworks, and provides flexible processing of enumeration types.